+++ firmware/src/drivers/user_interface.h
@@ -122,7 +122,7 @@ PUBLIC void
 user_interface_reset_tracking_target();
 
 PUBLIC void
-user_interface_set_movement_data( uint8_t move_id, uint8_t move_type, uint8_t progress );
+user_interface_set_movement_data(uint16_t move_id, uint8_t move_type, uint8_t progress );
 
 PUBLIC void
 user_interface_set_pathing_status( uint8_t status );
